refactor index updating logic into separate module
no code changes
